Fix PHP 8 compatibility: cast equipment_id to int
GETPOST('equipment_id', 'int') returns a string in Dolibarr, never an int.
In PHP 8, comparing a non-numeric string to 0 uses string comparison,
so "" >= 0 evaluates to FALSE (unlike PHP 7 where both were cast to 0).
This caused the save_entry action to silently fail for general entries
(Allgemeine Arbeiten) when equipment_id was not present in the URL.

Fix: (int)GETPOST('equipment_id', 'int') ensures equipment_id is always
an integer (0 when absent/empty), so >= 0 correctly evaluates to TRUE.

+++ intervention_equipment_details.php
@@ -36,7 +36,7 @@ $langs->loadLangs(array('interventions', 'equipmentmanager@equipmentmanager', 'p
 $id = GETPOST('id', 'int');
 $ref = GETPOST('ref', 'alpha');
 $action = GETPOST('action', 'aZ09');
-$equipment_id = GETPOST('equipment_id', 'int');
+$equipment_id = (int)GETPOST('equipment_id', 'int');
 $material_id = GETPOST('material_id', 'int');
 $entry_id = GETPOST('entry_id', 'int');
